# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FlexPos.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5088/"
    }
  ],
  "paths": {
    "/api/autenticacion/login": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InicioSesionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InicioSesionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/InicioSesionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/autenticacion/refrescar": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/autenticacion/logout": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/autenticacion/cambiar-contrasena": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarContrasenaRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarContrasenaRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarContrasenaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaAcceso"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CambiarContrasenaRequest": {
        "required": [
          "contrasenaActual",
          "contrasenaNueva"
        ],
        "type": "object",
        "properties": {
          "contrasenaActual": {
            "type": [
              "null",
              "string"
            ]
          },
          "contrasenaNueva": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "InicioSesionRequest": {
        "required": [
          "correo",
          "contrasena"
        ],
        "type": "object",
        "properties": {
          "correo": {
            "type": [
              "null",
              "string"
            ]
          },
          "contrasena": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RespuestaAcceso": {
        "required": [
          "tokenAcceso",
          "expiraUtc",
          "segundosVigencia"
        ],
        "type": "object",
        "properties": {
          "tokenAcceso": {
            "type": "string"
          },
          "expiraUtc": {
            "type": "string",
            "format": "date-time"
          },
          "segundosVigencia": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}